requirement.txt
appdirs==1.4.0
Django==1.10.6
django-hashers-passlib==0.1
django-widget-tweaks==1.4.1
djangorestframework==3.5.4
MySQL-python==1.2.5
packaging==16.8
passlib==1.6.2
pyparsing==2.1.10
six==1.10.0
